make: do not claim we wrote something if we didn't
If the build failed, then print nothing.

If the build succeeded either because we did a great job or because
there was nothing to be done, then print the list of relevant files
as before.  But prefix that with "Success" instead of "Wrote" because
in the later case we did not actually write those files just now.

Also put the "Success" message on a separate line so that the first
file aligns with the others.